feat: Add velodrome referral event filtering
