+++ scripts/calculateRevenue/protocols/velodrome/constants.ts
@@ -6,3 +6,5 @@ export const VELODROME_SUPPORTED_LIQUIDITY_POOL_ADDRESSES: Address[] = [
 ]
 
 export const VELODROME_NETWORK_ID = NetworkId['op-mainnet']
+
+export const VELODROME_UNIVERSAL_ROUTER_ADDRESS = '0x4bF3E32de155359D1D75e8B474b66848221142fc'
