+++ scripts/protocolFilters/velodrome.ts
@@ -1,6 +1,11 @@
+import { VELODROME_NETWORK_ID, VELODROME_UNIVERSAL_ROUTER_ADDRESS } from '../calculateRevenue/protocols/velodrome/constants'
 import { ReferralEvent } from '../types'
+import { filterDrome } from '../utils/filterDrome'
 
-// TODO (ENG-270): Add in same filtering as with aerodrome
 export async function filter(event: ReferralEvent): Promise<boolean> {
-  return !!event
+  return filterDrome({
+    event,
+    routerAddress: VELODROME_UNIVERSAL_ROUTER_ADDRESS,
+    networkId: VELODROME_NETWORK_ID,
+  })
 }
